Add benchmarks for mapping passes
This commit adds new benchmark methods for mapping passes. There are 2
benchmarks for each analysis pass that track the run time of the pass
and the memory consumption. Then for transformation passes a third
benchmark is added to track the depth after the pass is run. For swap
mappers a 4th benchmark also keeping track of how many swap gates are
in the circuit.