Add benchmarks for mapping passes (#699)
* Add benchmarks for mapping passes

This commit adds new benchmark methods for mapping passes. There are 2
benchmarks for each analysis pass that track the run time of the pass
and the memory consumption. Then for transformation passes a third
benchmark is added to track the depth after the pass is run. For swap
mappers a 4th benchmark also keeping track of how many swap gates are
in the circuit.

* Fix lint

The linter fails because two passes have not been released yet, this
commit disables this check for the file so we can run the benchmark for
these new passes.

* Add python-constraint to asv venv

The csp layout pass uses python-constraint which is an optional
dependency. This needs to be manually installed to make the csp layout
pass benchmarks work, this commit adds it to the list of dependencies
installed.