Install seaborn in asv venv (#712)
Qiskit/qiskit-terra#3508 added an inadvertant hard coded import-time
dependency on having seaborn installed if matplotlib is present even
though we do not use the visualization functions in benchmarking. On
the benchmarking machine used to generate results at
https://qiskit.github.io/qiskit/ matplotlib is present on the system
which causes a ImportError traceback for all commits after
Qiskit/qiskit-terra#3508. A fix for this has been proposed in
Qiskit/qiskit-terra#3538 but to benchmark the commits between
Qiskit/qiskit-terra#3508 and when Qiskit/qiskit-terra#3538 merges we'll
need to have seaborn present in the asv built venv. This commit adds
seaborn to the asv install config to make sure we generate data for
those commits.

Related to Qiskit/qiskit-terra#3525