Fix lint and docs build
This commit fixes the docs lint errors reported by doc8. At the same
time it fixes the requiremetns for the docs build with new terra by
installing the visualization requirements for qiskit in the docs venv.
The docs now are dynamically generating visualizations as part of the
build and to make this work we need to have all the dependencies
installed.