Update AUTHORS list and bibtex file
We recently pushed the 0.14.0 release out, but neglected to update the
authors list as part of that. The qiskit-bot automation is supposed to
take care of this, but it encountered a bug during the release process
and timed out. This commit fixes that oversight and updates the authors
list by running ./tools/generate_authors.py and then updates the bibtex
file by running ./tools/generate_bibtex.py.