Add Spanish from Spain
I'd like to contribute by translating to Spanish from Spain. I've been living here since the age of 5 and Spanish is basically as much of a native language for me as English is. I'd also love to eventually contribute to code as I learn more about Quantum Computing as a whole, but for now I believe this is the best I can give.

Regards,
Brandon Solo