Update python package long_description to use readme
Right now the meta-package's description page is a short string which is
hardcoded in the setup.py. This string is also currently out of date
with the recent release of 0.15.0 which included python 3.8 support
(although the trove classifier correctly indicates this). To provide
more information about the package this changes the package description
to be the README's contents. It will be less likely to go stale since
there's only a single place to update and it will give the same view on
pypi as we have on github.