Fix docs from master build
Passing all the requirements to a single pip command was error prone and
unreliable. To address this a separate pip command is used for
installing each element from github. This also adjusts the order the
elements are installed to ensure we upgrade terra first since it is the
dependency of all the other elements.