Update test configuration (#804)
Now that we're supporting python 3.8 on releases now we should be
verifying that everything works on that environment too. This commit
updates the job configuration to also include a python 3.8 job. It also
takes the opportunity to cleanup the travis config, remove some
deprecated fields, and standardize the environment on newer versions of
system libraries.