Update package metadata in setup.py (#803)
* Update python package long_description to use readme

Right now the meta-package's description page is a short string which is
hardcoded in the setup.py. This string is also currently out of date
with the recent release of 0.15.0 which included python 3.8 support
(although the trove classifier correctly indicates this). To provide
more information about the package this changes the package description
to be the README's contents. It will be less likely to go stale since
there's only a single place to update and it will give the same view on
pypi as we have on github.

* Add project urls to package metadata

To add useful links on the sidebar of the pypi page for the project this
commit adds project urls to the package metadata. On the next qiskit
meta-package release individual links to the source code, bug tracker,
and rendered documentation will be added to the pypi package page.

* Add missing README variable

* Add missing import

Co-authored-by: Kevin Krsulich <kevin@krsulich.net>