Explicitly call load_api_sources()
Having load_api_sources() run as part of the sphinx app api state
machine in the build-inited phase was too late. This was preventing the
stubs from being generated as part of a normal build causing all the
docs builds to fail for everything. This commit fixes that by explicitly
calling the function as the first step in the setup() function for the
sphinx app. This means that copying the sources over from the elements
is always the first thing we do as part of a sphinx build now.