Fix typo in deploy_documentation.sh
In #817 we added a step to sync setup.py to poBranch as part of
documentation deploy (as part of the en po file updateds. However, that
commit had a typo which prevented it from working. That being said it
was actually a good thing, because setup.py was added to the wrong spot
in the script. If the typo wasn't there we'd have run git rm setup.py
and removed it from poBranch. This commit fixes this issue (and the typo
by extension) so that we only do a cp of setup.py instead of removing
it.