Bump numpy on poBranch (#822)
* Bump numpy on poBranch

The translated docs builds are failing because of Qiskit/qiskit-aer#592
which is caused by using a version of numpy incompatible with the pulse
simulator. This has been fixed in aer on master and the stable/0.4.0
branch pending a 0.4.1 release. But, until that is ready we need to
manual ensure we're using a newer numpy version to avoid this issue in
the docs builds.

* Bump pip and friends too