Stop showing verbose rclone progress and use travis wait (#842)
The verbose output from using --progress with rclone uploads is hitting
the upper log size limit for large docs changes. This was originally
added in #807 so that we didn't have travis stop the CI job because the
upload took >10mins (with no output). To workaround both issues this
commit uses travis_wait and removes the --progress flag from rclone.

Co-authored-by: Paul Nation <nonhermitian@gmail.com>