Integrate Aer benchmarks within Qiskit metapackage
With the goal in mind of having all regression benchmarks in one central point, all
of the AER benchmarks have been moved away from Aer repo to the Qiskit metapackage
repository, where Terra benchmarks reside at the moment. Because there's a lot of
code in common between Terra and Aer benchmarkings, I have refactored a bit of code
as part of the integration to gain in clarity and consistency.
Aer benchmarks will run in their own baremetal instance though, to avoid interferences
with Terra benchmarkings, but the results of the benchmark will be published in the
same place.