Add travis job to run pip check (#864)
* Add travis job to run pip check

The new pip depsolver is coming this summer. [1] Until it's ready
this starts running pip check as a travis job to make sure we're
not somehow installing conflicting package versions. We shouldn't
be, but if we are better to know and start preparing/fixing it
now.

[1] https://pyfound.blogspot.com/2020/03/new-pip-resolver-to-roll-out-this-year.html

* Integrate pip check into existing jobs

This commit removes the standalone pip check job and instead integrates
it into the standard unit test jobs. This should save time since we
don't need to wait for an additional travis node just to run pip check.