Install pygments via asv to unblock benchmark runs (#867)
In Qiskit/qiskit-terra#3861 a new requirement was added to terra without
updating the requirements list in the setup.py. This causes issues when
trying to run the benchmarks for commits after that PR and prior to
Qiskit/qiskit-terra#4045 merging because it can't actually install
terra. This commit adds pygments to the asv.conf.json to ensure we can
have continuous benchmarking.