Workaround terra backwards compat bug
This commit works around an issue introduce in the most recent terra
release where the deprecated gate classes are no longer being exported,
see Qiskit/qiskit-terra#4365 for more details. This will be fixed soon
in a followup 0.14.1 release but until that time this commit makes the
necessary change to the benchmarks so they'll work with the 0.14.0
release.