Increase travis wait timeout and enable parallel builds (#910)
The combined doc job is taking over 45 mins to build and upload
currently. While the docs builds finishes the job exits because of a
timeout while uploading the new files to object storage. This commit
attempts to mitigate this by increasing the travis_wait timeout to 50min
which is the maximum allowed by travis. Also in an attempt to build the
documentation faster the sphinx-build -j flag is used to parallelize the
build across both vCPUs on the ci nodes. This should hopefully provide
enough of a margin to build the documentation.