Remove typehints plugin to try and fix deprecation warnings
This commit removes the autodoc typehints module in an attempt to fix
the deprecation warnings on sphinx >3.0.0 that are raised by every
function signature that uses type hints. Sphinx is supposed to have
native support for typehints this commit is to see if the output from
the docs builds with type hints are formatted decently without the
plugin and to see if the type hints go away.