Add minimum numpy version to requirements-dev (#921)
This commit adds a minimum numpy version to the requirements list to fix
the translated docs builds with qiskit>=0.18.3. Since Aer 0.5.1 the
minimum numpy version hasn't been satisifed and because of pip's lack of
a dependency solver it's unable to install the newer version that's
conflicting with the pre-installed system python numpy version. This
commit adds an explict requirement on the minimum numpy version to
ensure that install a new enough version prior to building the docs.
At the same time the version of pip, setuptools, etc are also updated
for the docs build jobs to ensure that we have a new enough version
installed.