Add sphinx extension to clone and built tutorials (#922)
This commit adds a new custome extension to the sphinx build's conf.py
that will clone master of the tutorials to the documentation and use
nbsphinx to build them as part of the documentation. Currently the
notebooks are not being executed and it will rely on the tutorials repo
including the cell output in the in-repo copy. What's missing from this
is an index and inclusion in the toc so while the tutorials are built
they aren't navigable and have to be manually opened.