Correct oversight in po sync job (#941)
In #938 the process of switching the repo where translations run to
qiskit-community/qiskit-translations was started by updating the deploy
script to push to that repo instead of poBranch on qiskit/qiskit.
However, in that change the script was updated to push the output to a
different repo but missed updating the actual cloning a working copy of
the dest repo. This commit corrects the oversight.