Add reverse CI job for building tutorials (#945)
The qiskit-tutorials repo is using the qiskit metapackage to install
qiskit and running the tutorials in CI. However, we do not have a job in
place to verify that a pending release will not break the tutorials.
Normally we would do this as part of a docs build but because of our
limited time budget we can't do a full doc build and tutorial build in
the same job. This commit just runs the tutorials sphinx build like it
does in the tutorials CI, but uses master from Qiskit/qiskit. This way
we'll catch any issues in a pending release prior to pushing the tag.