Unblock docs publish job (#951)
In the recent ignis 0.3.2 (and 0.3.1, but that had other bugs) release a
bug slipped in that crashes ignis if cvxpy is not installed. This is
causing the docs jobs to fail to build any documentation that ends up
importing process tomography (see Qiskit/qiskit-ignis#429). This commit
explicitly adds cvxpy to the docs jobs to unblock them so we upload
complete documentation. An ignis bugfix release 0.3.3 will be out soon
to fix the ignis bug, but until then this will fix the hosted
documentation.