Remove third-party providers from generate_authors.py (#940)
The qiskit repo is in the process of being reorganized to concentrate on
the core elements, dependencies, development tools, etc. As part of this
reoganization 3rd party providers are being moved to projects under
qiskit-community. This commit removes the 3rd party providers from the
list of projects that are used in the generate authors script to mirror
this change.

Co-authored-by: Paul Nation <nonhermitian@gmail.com>