Stop relying on external links in the_elements.rst (#970)
* Stop relying on external links in the_elements.rst

This commit fixes the_elements.rst to stop relying on external links for
links to api documentation. These are elements in the documentation that
should use internal links. Mostly because the pieces can move around and
result in dead links. Additionally, for local docs builds this will
result in going to the hosted documentation instead of the local builds.
We should never be relying on external links to the documentation from
inside the documentation.

Fixes #969

* Fix typo

* Update docs/the_elements.rst

Co-authored-by: SooluThomas <soolu.elto@gmail.com>

Co-authored-by: SooluThomas <soolu.elto@gmail.com>