Cap matplotlib and add set -e to deploy script (#973)
The recent matplotlib broke the hinton plotter (which has been fixed on
terra master) and the documentation fails to build now when it runs the
docs for the hinton plot function. Because the deploy script was missing
a set -e the failed documentation build didn't stop the script and it
continued to run a sync on empty directory which lead to the
documentation getting deleted. This commit fixes the docs builds by
pinning matplotlib until the next terra release and setting set -e on
the deploy script so we the documentation doesn't accidently get deleted
when the docs build fails.