Fix the tutorials job (#974)
* Fix the tutorials job

The recent release of matplotlib 3.3.0 broke the plot_state_hinton
visualization function (see Qiskit/qiskit-terra#4739 for more details).
This has resulted in docs builds and tutorials using released version of
qiskit-terra to fail when they go to use plot_state_hinton. This broke
the docs builds and end up wiping out the documentation see #973. While
that has been fixed for documentation the tutorials build jobs are still
broken because they do not cap matplotlib. This commit adds a cap on
matplotlib to unbreak the job. When the next qiskit-terra release comes
out we can remove the version cap.

* Add quotes to matplotlib cap