Add pyscf to requirements-dev.txt
The latest qiskit meta-package release made pyscf optional. However it
is still needed to build the chemistry docs. #998 added it to the travis
config for the tutorials job but neglected to update the config for doc
publishing to manually install pyscf. This commit corrects the oversight
to unblock the docs builds.