Increase nbsphinx timeout to 5 mins (#1007)
Prior to #989 we were just deploying prebuilt tutorials in the docs
publish jobs so the per cell timeout was set to 60sec which should have
been plenty since nothing was executed. However after #989 in the docs
deploy job we started executing all the cells. The 60 sec timeout is not
sufficient for that, especially considering the tutorials repo has a
3min per cell timeout. This commit bumps the timeout to 5min to give a
healthy margin on the publish job, where there isn't really any time
pressure.