Remove CHANGELOG.md file (#996)
The changelog file hasn't been used for some time, as the contents of it
have moved to the release pages since the adoption of qiskit-bot to
automate version bumps. Having the file in tree is confusing since it is
out of date and contains duplicate information. This commit removes the
file to avoid this moving forward.

Co-authored-by: SooluThomas <soolu.elto@gmail.com>