Add back translatable string job (#1051)
In #1038 we migrated the Qiskit metapackage CI to use Github Actions
instead of Travis CI. However, as part of that migration one job was
overlooked and not migrated from Travis to Github Actions, the
translatable string post-merge job which generates the po files and
updates the qiskit-community/qiskit-translations repo which new
translatable strings in the docs. The translation repo hasn't
received string updates since we migrated to github actions because
in #1038 we accidently deleted the job and didn't add an equivalent
to Github Actions. This commit corrects the oversight and adds back
a translatable string job to Github Actions.