Remove peakmem benchmarks from the benchmark suite
This commit removes all the peakmem benchmarks from the benchmark suite.
The peakmem benchmarks have proven to be unreliable in practice and are
prone to inaccurately flag large regressions in the amount of memory
used (see Qiskit/qiskit-terra#4689 and Qiskit/qiskit-terra#5247) which
are unreproduceable and also don't match what running processes on the
benchmark show. This might be an artifact of the asv launch method used,
but either way if we can't rely on the benchmarks to provide a
consistent baseline running benchmarks to measure memory consumption is
a waste of CPU time that could be spent providing wider benchmarking
coverage.