Compress docs and tutorials artifacts before upload (#1088)
* Compress docs and tutorials artifacts before upload

This commit compresses the docs build artifacts before we upload it. The
output from the docs and tutorials jobs are quite large both in space
and number of files. This causes the upload job to bog down and take a
long time while it iterates over each file. By compressing it prior to
upload it should decrease this overhead since there is a single to
upload.

* Actually put compressed file in artifacts dir