Bump cell timeout in the tutorials CI job
The 07_asian_barrier_spread_pricing tutorial is quite slow and often
times out in CI because running the simulation takes very close to 3
minutes in the fastest case. But, the performance of the CI nodes is
quite variable since they run on virtual machines in azure's public
cloud. So many jobs fail randomly because of timeouts on that notebook.
In Qiskit/qiskit-tutorial#1091 we added an env var to adjust the per
cell timeout for CI jobs. This commit leverages that env var in an
attempt to fix the timeout issues by bumping the timeout to 5 min which
should hopefully give us a healthy margin for slow CI nodes. Especially
since the upstream tutorials repo CI is still holding incoming PRs to 3
mins per cell.