Pin scipy version in CI (#1139)
The recent release of scipy 1.6.0 is causing the qsvm tutorials to fail
(see Qiskit/qiskit-aqua#1498 for more details). This is blocking the
docs deploy CI job because it runs the tutorials (as well as the likely
the tutorials job too) and preventing us from uploading new versions of
the built documentation. This commit pins the scipy version to a known
working version (1.5.4) to unblock CI until we have a fix for the
tutorial failure with scipy 1.6.0.