Fix the Pulse schedule construction benchmarks (#1147)
The Pulse schedule construction benchmarks were using the previously
deprecated and now removed SamplePulse class as the pulse to add to a
schedule. This has been failing in benchmark runs since that class was
removed. This commit fixes the failure by using the replacement class
Waveform. This will break compatibility with old versions of terra from
before waveform existed, however since we're unlikely to run a backfill
job for historical data this is an ok tradeoff.