Mark versionutils extension as parallel_read_safe
The versionutils sphinx extension we use during the combined
documentation is raising a warning on each sphinx build about it not
specifying a parallel_read_safe option. This warning while not fatal has
2 issues, first it is defaulting to False which means we're not
parallelizing the build where versionutils is used (which is in 1 place
so it's not crticial) the second is that if/when we're able to set -W
during builds this would be a blocker. This commit returns the
parallel_read_safe option as True since it should be safe to have the
extension run concurrently. It relies on git history and readonly  git
commands which are external sources of truth and won't be affected by
parallel access (or any internal git lock files). This should eliminate
the warning and enable parallelization where the extension is used.