Exclude other documentation builds from rclone sync
In the near future there will be several documentation builds for qiskit
projects hosted off the https://qiskit.org/documentation page that will
be self contained in the project repository and uploaded directly by the
project. To avoid the main qiskit doc uploads for deleting that content
this commit adds an exclude list for all those other builds. This
includes the existing locale build directory for translated documentation
which we were already excluding.