Remove benchmarks that measure BasicAer as transpile target
This commmit removes all the benchmarks which were using the basic aer
simulator as a target for transpilation. These benchmarks were
originally added in an attempt to show the overhead of using routing and
layout in a transpile but never did a good job of this. We have better
microbenchmarks now of each individual transpiler pass and the basic aer
target isn't a realistic use case for anything anymore. By removing
these duplicate benchmarks we get back some of our time budget that we
can use for adding new benchmarks.

There are 2 benchmarks which aren't updated to remove the basic aer
usage. The first is the quantum_volume module which is addressed in
PR #1185 as part of a larger refactor of that benchmark and the state
tomography benchmarks which are are actually running a simulation in
basic aer and its sill necessary there.