Update QuantumVolume benchmark (#1185)
* Update QuantumVolume benchmark

This commit updates the quantum volume benchmark to be a bit more modern
and more closely represent a typical quantum volume use cases. Instead
of varying the width and depth of the input circuit as parameters it
uses a sweep over just width (and depth matches) and the depth matches
to actually use circuits people run and also to reduce the number of
permutations. Additionally, it updates the target device to use a more
recent ibmq device (paris) which changes the target coupling map and
basis gates. The basic aer target is removed from the benchmark suite
because it provides limited value as it's basically running with
no coupling map and the basis gates u1, u2, u3 which isn't a realistic
target.

At the same time this drops the backwards comaptibility shims used to
run the benchmark against both old and new versions of qiskit and
instead relies on current methods of generation and is only concerned
with the performance moving forward.

* Fix lint

* Fix pylint