Add benchmarks using Qiskit aer as a target
In a follow up to #1186 where we removed all the benchmarks using
BasicAer as a compilation target this commit adds some new benchmarks
that use qiskit-aer as a compilation target. Mainly a single qv
circuit (at all 4 optimization levels) and randomized benchmarking (
both single process and parallel). The state tomography benchmarks were
also updated to use aer instead of basic aer, however those benchmarks
also run a simulation so it's more than just a compilation target
difference. This use of aer is done using a pinned version for stability
in the results when we change the aer version the benchmarks using it
will have to be updated too so we know the results are associated with
a single aer version.

Since we had to invalidate the results/bump the version of the rb
benchmarks this was a good opportunity to update the ignis version to
the latest release (which has a similar pinned version arrangement).