Cap tweedledum
With the recent release of tweedledum 1.0.0 the docs builds have started
to fail. This is because the released versions of qiskit-terra (0.16.x)
are incompatible with that new tweedledum release. The next terra
release will fix compatibility (and bump the minimum version), but until
terra is released this commit caps the tweedledum version.

Fixes #1202