Enable using version table extension outside of qiskit repo (#1174)
Fixes #1008

This fixes the issue by downloading the Qiskit Metapackage into a tmp folder. git tag for translated docs
returns an empty list in "Qiskit Metapackage Version" because translated docs are in
qiskit-community/qiskit-translations. Instead of doing that, a new function has been added, It downloads
the git repo into a temporary folder and then the git tag command is executed so it always return the
latest version list from the official repo.