Add cplex to docs builds (#1211)
The aqua migration guide that documents how to migrate from using aqua
to the new application packages and terra instead is a jupyter notebook
and requires cplex to execute it. For the publish docs job we run the
notebooks to ensure a consistent output in the cells, but were not
installing cplex. This is causing the docs publish jobs to fail when it
tries to execute the migration notebook because the dependency is
missing. This commit fixes this by just adding cplex to the docs job tox
venv so that it's installed. We should remove this when aqua goes eol
and we no longer need to provide a migration guide.