Disable notebook execute in docs publish job temporarily
Recently the docs tutorials included in the published documentation
builds have not been rendered correctly and are missing content. While
the output from the docs CI job look fine. The only functional
difference between the jobs is the published docs are set to execute the
notebooks while the CI job doesn't (in the interest of time, and we have
a dedicated job which just builds the tutorials separately for testing)
and instead just shows the included cell output. To attempt to address
the broken builds this commit disables the environment variable used to
trigger executing the notebook for the publish job. This will hopefully
fix the published version of the docs until we can get to the bottom of
what has changed and is breaking the tutorials in the published builds.