Pin docutils for docs builds
The recent release of docutils 0.17 has started to break the tutorials
output when we execute the notebooks. To workaround this we previously
disabled executing the notebooks for the publish job in #1225. However,
manually testing shows that it was the docutils release and this commit
reverts #1225 and pins docutils to unbreak the build.