Add missing Terra 0.17.1 release notes
As part of the Terra 0.17.1 release prep a release note file was
accidentally not committed to the tree. This meant when the rst was
generated from reno after the release these release notes were missing
and not copied into the release_notes.rst file. This commit manually
copies the missing release notes into the document so they're there for
the historical record.