Add support for stable versions of documentation (#1261)
* Add support for stable versions of documentation

This commit adds support for hosting stable versions of the
documentation moving forward. This is accomplished by adding a new ci
workflow on tags that will upload docs to stable/$VERSION where $VERSION
is the first 2 components ('major.minor') from the version string. This
will let us maintain a hosted copy of each minor release of
documentation for users still running on older versions. The version
template and custom sphinx extensions are updated to determine the list
of versions from the current state of the git repo.

This PR assumes that we've manually updated 0.24.x, 0.25.x, and 0.26.x
to 'documentation/stable/0.24', 'documentation/stable/0.25', and
'documentation/stable/0.26' respectively. This is a blocker for merging
this because until that done the version menu will have broken links.

* Fix typo in template

* Update docs/_templates/versions.html

* Fix source path in rclone upload

* Apply suggestions from code review

Co-authored-by: SooluThomas <soolu.elto@gmail.com>

Co-authored-by: SooluThomas <soolu.elto@gmail.com>