Add redirects for application tutorials
The tutorials for the application modules have been moved out of the
qiskit-tutorials repository to the application module documentation.
This commit adds the approriate redirects so existing links work as
expected. It also updates the tutorials index and requirements list
to remove things that were only there for the application tutorials.