Add back cplex to docs job
In the recently merged #1266 we removed the cplex requirement from the
tox docs job because we removed the application module tutorials which
were dependent on this. However, that neglected to take into account the
aqua migration guide tutorial which we're still going to host and it
requires cplex to execute. Removing it has caused the docs builds to
fail on the publish job which is the only place we execute the migration
tutorial. This commit adds cplex back to fix the job and unblock docs
updates.