Fix generate authors script
The generate authors script has been failing when qiskit-bot goes to
update it because there was an invalidly formatted co-authorship line
in one of the commit messages. This was causing the parser to fail
because it couldn't find an email address for the co-author. This commit
updates the script so if that occurs it just skips that line.